Create README.md
New to the world of coding currently self teaching. Only able to work with Tremux AND all it's associated tools on android.  Highly motivated to learn as much as possible .  Self teaching. Good understanding of basic fundamentals of computers and software development. As well as Termux commands. Very quick learner.
Not here to waste anyone's time just motivated to learn. 
Look forward to the future